+++ docs/fr_FR/Changelog.rst
@@ -4,6 +4,8 @@ ChangeLog
 - Girier curtain module: Mise-à-jour du modèle (2526).
 - Interne: Parser: Correction warning PHP 'Binding table response'.
 - Modele TS201 renommage automatique vers 'TS0201__TYZB01_hjsgdkfl'.
+- Page avancé: Support préliminaire 'unbind'.
+- Interne: Cmd: Support préliminaire 'unbind0031'.
 
 230320-BETA-3
 -------------